Implement further allocation methods
Implemented a lot of the allocation methods. Not doing the `try` ones,
as they seem to assume memory allocation is fallible ‒ but, is it?

Also, not implementing a lot of the statistical/querying methods. They
don't really make sense when the `Bump` can be reused further down the
line and the allocations are spread across multiple arenas. It could
make sense on the `Herd` itself, but it would not allow us to take the
members out completely.